Initial commit
My emulator project.  As of now (and probably until the end of the days), just
a 6502 emulator.  The idea is to have separate CPU/PPU modules with skinny APIs
and interface them with a main binary that sets up the hooks between the 2,
ideally allowing different PPUs to be attached to the CPU (right now, only the
6502).  Note that no PPUs are emulated yet, (look for a directory called ppu
in the future :P)

Different PPUs to be attached, apart from the NES, include the Atari 2600 and
the 6502 emulator from http://6502asm.com