Optimizations on simple6502
Refresh the framebuffer outside of the CPU hook.  Performance is decent but
can be improved by removing the usleep() at step_cb() time.  Improved means
that you don't even see whats going on, so let's keep it with the usleep().

In the future it might be nice to do some measurements to give a constant
framerate in every CPU this might run on.