NES: Implement PPU Memory and Control Registers
Unfortunately there is little proof that this works at all.  ROMS execute and
go in a loop (probably waiting for other unimplemented subsystems), but it's
difficult to the see if this is working without getting the full blown
graphics.  The "reference" documentation for this is the classic NES
Architecture by Marat Fayzullin:
 * http://fms.komkon.org/EMUL8/NES.html