Different memhook types can have different different callback signature
Memhooks used to be so they were sharing the callback signature.  This was not
so clean for read/write hooks (since read was simply ignoring part of the
signature).  Given that we need a different type of hook for mirroring address
space, which doesn't adapt to the existing common signature, this change was a
logical first step.

memhook_fun_t is now a union of different signatures.  CPU users have no impact
other than a different hook setup set of functions.  These functions no longer
pass a mh_type_t enum but the type is encoded in the function name.