BUILD: Rewrite of the build system
...until is not even recognizable.
This rewrite tries to do with implicit rules, so the main logic resides in the
included .mk files.  The parts were no refactoring makes sense are more or less
intact (for instance, there's no reason to move the coverage to the .mk's).

The idea is that Makefiles are as shorter as possible and made of definitions
that the included rules have to chew.  This revamp actually contains some fancy
new facilities while the amount of lines of the build system remains
acceptable.